+++ controller/konnect/ops/ops_controlplane.go
@@ -3,11 +3,19 @@ package ops
 import (
 	"context"
 	"errors"
+	"fmt"
+	"sort"
 
 	sdkkonnectgo "github.com/Kong/sdk-konnect-go"
 	sdkkonnectcomp "github.com/Kong/sdk-konnect-go/models/components"
 	sdkkonnecterrs "github.com/Kong/sdk-konnect-go/models/sdkerrors"
+	"github.com/samber/lo"
+	"github.com/sourcegraph/conc/iter"
+	corev1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/kong/gateway-operator/controller/konnect/conditions"
 
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
@@ -15,6 +23,8 @@ import (
 func createControlPlane(
 	ctx context.Context,
 	sdk ControlPlaneSDK,
+	sdkGroups ControlPlaneGroupSDK,
+	cl client.Client,
 	cp *konnectv1alpha1.KonnectGatewayControlPlane,
 ) error {
 	req := cp.Spec.CreateControlPlaneRequest
@@ -30,6 +40,12 @@ func createControlPlane(
 	}
 
 	cp.Status.SetKonnectID(*resp.ControlPlane.ID)
+
+	if err := setGroupMembers(ctx, cl, cp, sdkGroups); err != nil {
+		SetKonnectEntityProgrammedConditionFalse(cp, conditions.KonnectGatewayControlPlaneProgrammedReasonFailedToSetControlPlaneGroupMembers, err.Error())
+		return err
+	}
+
 	SetKonnectEntityProgrammedCondition(cp)
 
 	return nil
@@ -75,6 +91,8 @@ func deleteControlPlane(
 func updateControlPlane(
 	ctx context.Context,
 	sdk ControlPlaneSDK,
+	sdkGroups ControlPlaneGroupSDK,
+	cl client.Client,
 	cp *konnectv1alpha1.KonnectGatewayControlPlane,
 ) error {
 	id := cp.GetKonnectStatus().GetKonnectID()
@@ -93,7 +111,7 @@ func updateControlPlane(
 			Info("entity not found in Konnect, trying to recreate",
 				"type", cp.GetTypeName(), "id", id,
 			)
-		if err := createControlPlane(ctx, sdk, cp); err != nil {
+		if err := createControlPlane(ctx, sdk, sdkGroups, cl, cp); err != nil {
 			return FailedKonnectOpError[konnectv1alpha1.KonnectGatewayControlPlane]{
 				Op:  UpdateOp,
 				Err: err,
@@ -113,7 +131,70 @@ func updateControlPlane(
 	}
 
 	cp.Status.SetKonnectID(*resp.ControlPlane.ID)
+
+	if err := setGroupMembers(ctx, cl, cp, sdkGroups); err != nil {
+		SetKonnectEntityProgrammedConditionFalse(cp, conditions.KonnectGatewayControlPlaneProgrammedReasonFailedToSetControlPlaneGroupMembers, err.Error())
+		return err
+	}
+
 	SetKonnectEntityProgrammedCondition(cp)
 
 	return nil
 }
+
+func setGroupMembers(
+	ctx context.Context,
+	cl client.Client,
+	cp *konnectv1alpha1.KonnectGatewayControlPlane,
+	sdkGroups ControlPlaneGroupSDK,
+) error {
+	if len(cp.Spec.Members) == 0 ||
+		cp.Spec.ClusterType == nil ||
+		*cp.Spec.ClusterType != sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup {
+		return nil
+	}
+
+	members, err := iter.MapErr(cp.Spec.Members,
+		func(member *corev1.LocalObjectReference) (sdkkonnectcomp.Members, error) {
+			var (
+				memberCP konnectv1alpha1.KonnectGatewayControlPlane
+				nn       = client.ObjectKey{
+					Namespace: cp.Namespace,
+					Name:      member.Name,
+				}
+			)
+			if err := cl.Get(ctx, nn, &memberCP); err != nil {
+				return sdkkonnectcomp.Members{},
+					fmt.Errorf("failed to get control plane group member %s: %w", member.Name, err)
+			}
+			if memberCP.GetKonnectID() == "" {
+				return sdkkonnectcomp.Members{},
+					fmt.Errorf("control plane group %s member %s has no Konnect ID", cp.Name, member.Name)
+			}
+			return sdkkonnectcomp.Members{
+				ID: lo.ToPtr(memberCP.GetKonnectID()),
+			}, nil
+		})
+	if err != nil {
+		return fmt.Errorf("failed to set group members, some members couldn't be found: %w", err)
+	}
+
+	sort.Sort(membersByID(members))
+	gm := sdkkonnectcomp.GroupMembership{
+		Members: members,
+	}
+	_, err = sdkGroups.PutControlPlanesIDGroupMemberships(ctx, cp.GetKonnectID(), &gm)
+	if err != nil {
+		return fmt.Errorf("failed to set members on control plane group %s: %w",
+			client.ObjectKeyFromObject(cp), err,
+		)
+	}
+
+	return nil
+}
+
+type membersByID []sdkkonnectcomp.Members
+
+func (m membersByID) Len() int           { return len(m) }
+func (m membersByID) Less(i, j int) bool { return *m[i].ID < *m[j].ID }
+func (m membersByID) Swap(i, j int)      { m[i], m[j] = m[j], m[i] }