+++ controller/konnect/ops/ops_controlplane_test.go
@@ -11,11 +11,16 @@ import (
 	"github.com/google/uuid"
 	"github.com/samber/lo"
 	"github.com/stretchr/testify/assert"
+	mock "github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8stypes "k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/kong/gateway-operator/controller/konnect/conditions"
+	"github.com/kong/gateway-operator/modules/manager/scheme"
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
 
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
@@ -25,14 +30,15 @@ func TestCreateControlPlane(t *testing.T) {
 	ctx := context.Background()
 	testCases := []struct {
 		name        string
-		mockCPPair  func(*testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane)
+		mockCPTuple func(*testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane)
 		expectedErr bool
 		assertions  func(*testing.T, *konnectv1alpha1.KonnectGatewayControlPlane)
 	}{
 		{
 			name: "success",
-			mockCPPair: func(t *testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
+			mockCPTuple: func(t *testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
 				sdk := NewMockControlPlaneSDK(t)
+				sdkGroups := NewMockControlPlaneGroupSDK(t)
 				cp := &konnectv1alpha1.KonnectGatewayControlPlane{
 					Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
 						CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
@@ -55,7 +61,7 @@ func TestCreateControlPlane(t *testing.T) {
 						nil,
 					)
 
-				return sdk, cp
+				return sdk, sdkGroups, cp
 			},
 			assertions: func(t *testing.T, cp *konnectv1alpha1.KonnectGatewayControlPlane) {
 				assert.Equal(t, "12345", cp.GetKonnectStatus().GetKonnectID())
@@ -68,8 +74,9 @@ func TestCreateControlPlane(t *testing.T) {
 		},
 		{
 			name: "fail",
-			mockCPPair: func(t *testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
+			mockCPTuple: func(t *testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
 				sdk := NewMockControlPlaneSDK(t)
+				sdkGroups := NewMockControlPlaneGroupSDK(t)
 				cp := &konnectv1alpha1.KonnectGatewayControlPlane{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "cp-1",
@@ -95,7 +102,7 @@ func TestCreateControlPlane(t *testing.T) {
 						},
 					)
 
-				return sdk, cp
+				return sdk, sdkGroups, cp
 			},
 			assertions: func(t *testing.T, cp *konnectv1alpha1.KonnectGatewayControlPlane) {
 				assert.Equal(t, "", cp.Status.GetKonnectID())
@@ -112,9 +119,10 @@ func TestCreateControlPlane(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			sdk, cp := tc.mockCPPair(t)
+			sdk, sdkGroups, cp := tc.mockCPTuple(t)
+			fakeClient := fake.NewClientBuilder().Build()
 
-			err := createControlPlane(ctx, sdk, cp)
+			err := createControlPlane(ctx, sdk, sdkGroups, fakeClient, cp)
 			tc.assertions(t, cp)
 
 			if tc.expectedErr {
@@ -259,14 +267,15 @@ func TestUpdateControlPlane(t *testing.T) {
 	ctx := context.Background()
 	testCases := []struct {
 		name        string
-		mockCPPair  func(*testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane)
+		mockCPTuple func(*testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane)
 		expectedErr bool
 		assertions  func(*testing.T, *konnectv1alpha1.KonnectGatewayControlPlane)
 	}{
 		{
 			name: "success",
-			mockCPPair: func(t *testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
+			mockCPTuple: func(t *testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
 				sdk := NewMockControlPlaneSDK(t)
+				sdkGroups := NewMockControlPlaneGroupSDK(t)
 				cp := &konnectv1alpha1.KonnectGatewayControlPlane{
 					Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
 						CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
@@ -299,7 +308,7 @@ func TestUpdateControlPlane(t *testing.T) {
 						nil,
 					)
 
-				return sdk, cp
+				return sdk, sdkGroups, cp
 			},
 			assertions: func(t *testing.T, cp *konnectv1alpha1.KonnectGatewayControlPlane) {
 				assert.Equal(t, "12345", cp.Status.GetKonnectID())
@@ -313,8 +322,9 @@ func TestUpdateControlPlane(t *testing.T) {
 		},
 		{
 			name: "fail",
-			mockCPPair: func(t *testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
+			mockCPTuple: func(t *testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
 				sdk := NewMockControlPlaneSDK(t)
+				sdkGroups := NewMockControlPlaneGroupSDK(t)
 				cp := &konnectv1alpha1.KonnectGatewayControlPlane{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "cp-1",
@@ -351,7 +361,7 @@ func TestUpdateControlPlane(t *testing.T) {
 						},
 					)
 
-				return sdk, cp
+				return sdk, sdkGroups, cp
 			},
 			assertions: func(t *testing.T, cp *konnectv1alpha1.KonnectGatewayControlPlane) {
 				assert.Equal(t, "12345", cp.Status.GetKonnectID())
@@ -366,8 +376,9 @@ func TestUpdateControlPlane(t *testing.T) {
 		},
 		{
 			name: "when not found then try to create",
-			mockCPPair: func(t *testing.T) (*MockControlPlaneSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
+			mockCPTuple: func(t *testing.T) (*MockControlPlaneSDK, *MockControlPlaneGroupSDK, *konnectv1alpha1.KonnectGatewayControlPlane) {
 				sdk := NewMockControlPlaneSDK(t)
+				sdkGroups := NewMockControlPlaneGroupSDK(t)
 				cp := &konnectv1alpha1.KonnectGatewayControlPlane{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "cp-1",
@@ -418,7 +429,7 @@ func TestUpdateControlPlane(t *testing.T) {
 						nil,
 					)
 
-				return sdk, cp
+				return sdk, sdkGroups, cp
 			},
 			assertions: func(t *testing.T, cp *konnectv1alpha1.KonnectGatewayControlPlane) {
 				assert.Equal(t, "12345", cp.Status.GetKonnectID())
@@ -434,9 +445,10 @@ func TestUpdateControlPlane(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			sdk, cp := tc.mockCPPair(t)
+			sdk, sdkGroups, cp := tc.mockCPTuple(t)
+			fakeClient := fake.NewClientBuilder().Build()
 
-			err := updateControlPlane(ctx, sdk, cp)
+			err := updateControlPlane(ctx, sdk, sdkGroups, fakeClient, cp)
 
 			if tc.assertions != nil {
 				tc.assertions(t, cp)
@@ -472,7 +484,9 @@ func TestCreateAndUpdateControlPlane_KubernetesMetadataConsistency(t *testing.T)
 				},
 			},
 		}
-		sdk = NewMockControlPlaneSDK(t)
+		sdk        = NewMockControlPlaneSDK(t)
+		sdkGroups  = NewMockControlPlaneGroupSDK(t)
+		fakeClient = fake.NewClientBuilder().Build()
 	)
 
 	t.Log("Triggering CreateControlPlane with expected labels")
@@ -495,7 +509,7 @@ func TestCreateAndUpdateControlPlane_KubernetesMetadataConsistency(t *testing.T)
 				ID: lo.ToPtr("12345"),
 			},
 		}, nil)
-	require.NoError(t, createControlPlane(ctx, sdk, cp))
+	require.NoError(t, createControlPlane(ctx, sdk, sdkGroups, fakeClient, cp))
 
 	t.Log("Triggering UpdateControlPlane with expected labels")
 	sdk.EXPECT().
@@ -508,5 +522,278 @@ func TestCreateAndUpdateControlPlane_KubernetesMetadataConsistency(t *testing.T)
 				ID: lo.ToPtr("12345"),
 			},
 		}, nil)
-	require.NoError(t, updateControlPlane(ctx, sdk, cp))
+	require.NoError(t, updateControlPlane(ctx, sdk, sdkGroups, fakeClient, cp))
+}
+
+func TestSetGroupMembers(t *testing.T) {
+	testcases := []struct {
+		name        string
+		group       *konnectv1alpha1.KonnectGatewayControlPlane
+		cps         []client.Object
+		sdk         func(t *testing.T) *MockControlPlaneGroupSDK
+		expectedErr bool
+	}{
+		{
+			name: "no members",
+			group: &konnectv1alpha1.KonnectGatewayControlPlane{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "cp-group",
+					Namespace: "default",
+				},
+				Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+					CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+						Name:        "cp-group",
+						ClusterType: lo.ToPtr(sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup),
+					},
+				},
+			},
+			sdk: func(t *testing.T) *MockControlPlaneGroupSDK {
+				sdk := NewMockControlPlaneGroupSDK(t)
+				return sdk
+			},
+		},
+		{
+			name: "1 member with Konnect Status ID",
+			group: &konnectv1alpha1.KonnectGatewayControlPlane{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "cp-group",
+					Namespace: "default",
+				},
+				Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+					CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+						Name:        "cp-group",
+						ClusterType: lo.ToPtr(sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup),
+					},
+					Members: []corev1.LocalObjectReference{
+						{
+							Name: "cp-1",
+						},
+					},
+				},
+				Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+					KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+						ID: "cpg-12345",
+					},
+				},
+			},
+			cps: []client.Object{
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-1",
+						Namespace: "default",
+					},
+					Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+						KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+							ID: "cp-12345",
+						},
+					},
+				},
+			},
+			sdk: func(t *testing.T) *MockControlPlaneGroupSDK {
+				sdk := NewMockControlPlaneGroupSDK(t)
+				sdk.EXPECT().
+					PutControlPlanesIDGroupMemberships(
+						mock.Anything,
+						"cpg-12345",
+						&sdkkonnectcomp.GroupMembership{
+							Members: []sdkkonnectcomp.Members{
+								{
+									ID: lo.ToPtr("cp-12345"),
+								},
+							},
+						},
+					).
+					Return(
+						&sdkkonnectops.PutControlPlanesIDGroupMembershipsResponse{},
+						nil,
+					)
+				return sdk
+			},
+		},
+		{
+			name: "1 member without Konnect Status ID",
+			group: &konnectv1alpha1.KonnectGatewayControlPlane{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "cp-group",
+					Namespace: "default",
+				},
+				Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+					CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+						Name:        "cp-group",
+						ClusterType: lo.ToPtr(sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup),
+					},
+					Members: []corev1.LocalObjectReference{
+						{
+							Name: "cp-1",
+						},
+					},
+				},
+				Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+					KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+						ID: "cpg-12345",
+					},
+				},
+			},
+			cps: []client.Object{
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-1",
+						Namespace: "default",
+					},
+					Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{},
+				},
+			},
+			sdk: func(t *testing.T) *MockControlPlaneGroupSDK {
+				sdk := NewMockControlPlaneGroupSDK(t)
+				return sdk
+			},
+			expectedErr: true,
+		},
+		{
+			name: "2 member with Konnect Status IDs",
+			group: &konnectv1alpha1.KonnectGatewayControlPlane{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "cp-group",
+					Namespace: "default",
+				},
+				Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+					CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+						Name:        "cp-group",
+						ClusterType: lo.ToPtr(sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup),
+					},
+					Members: []corev1.LocalObjectReference{
+						{
+							Name: "cp-1",
+						},
+						{
+							Name: "cp-2",
+						},
+					},
+				},
+				Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+					KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+						ID: "cpg-12345",
+					},
+				},
+			},
+			cps: []client.Object{
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-1",
+						Namespace: "default",
+					},
+					Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+						KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+							ID: "cp-12345",
+						},
+					},
+				},
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-2",
+						Namespace: "default",
+					},
+					Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+						KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+							ID: "cp-12346",
+						},
+					},
+				},
+			},
+			sdk: func(t *testing.T) *MockControlPlaneGroupSDK {
+				sdk := NewMockControlPlaneGroupSDK(t)
+				sdk.EXPECT().
+					PutControlPlanesIDGroupMemberships(
+						mock.Anything,
+						"cpg-12345",
+						&sdkkonnectcomp.GroupMembership{
+							Members: []sdkkonnectcomp.Members{
+								{
+									ID: lo.ToPtr("cp-12345"),
+								},
+								{
+									ID: lo.ToPtr("cp-12346"),
+								},
+							},
+						},
+					).
+					Return(
+						&sdkkonnectops.PutControlPlanesIDGroupMembershipsResponse{},
+						nil,
+					)
+				return sdk
+			},
+		},
+		{
+			name: "2 member, 1 with Konnect Status IDs, 1 without it",
+			group: &konnectv1alpha1.KonnectGatewayControlPlane{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "cp-group",
+					Namespace: "default",
+				},
+				Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+					CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+						Name:        "cp-group",
+						ClusterType: lo.ToPtr(sdkkonnectcomp.ClusterTypeClusterTypeControlPlaneGroup),
+					},
+					Members: []corev1.LocalObjectReference{
+						{
+							Name: "cp-1",
+						},
+						{
+							Name: "cp-2",
+						},
+					},
+				},
+				Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+					KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+						ID: "cpg-12345",
+					},
+				},
+			},
+			cps: []client.Object{
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-1",
+						Namespace: "default",
+					},
+					Status: konnectv1alpha1.KonnectGatewayControlPlaneStatus{
+						KonnectEntityStatus: konnectv1alpha1.KonnectEntityStatus{
+							ID: "cp-12345",
+						},
+					},
+				},
+				&konnectv1alpha1.KonnectGatewayControlPlane{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "cp-2",
+						Namespace: "default",
+					},
+				},
+			},
+			sdk: func(t *testing.T) *MockControlPlaneGroupSDK {
+				sdk := NewMockControlPlaneGroupSDK(t)
+				return sdk
+			},
+			expectedErr: true,
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			fakeClient := fake.NewClientBuilder().
+				WithScheme(scheme.Get()).
+				WithObjects(tc.group).
+				WithObjects(tc.cps...).
+				Build()
+
+			sdk := tc.sdk(t)
+			err := setGroupMembers(context.Background(), fakeClient, tc.group, sdk)
+			if tc.expectedErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.True(t, sdk.AssertExpectations(t))
+		})
+	}
 }